ci(coverage): route per-cell reports into groups instead of merging them
The first cut merged each group's LCOV locally, reimplementing istanbul's
coverage-map merge as ~350 lines of LCOV text parsing. Both Codecov and Datadog
already merge same-flag uploads server-side — that is how they combined the ~430
per-cell reports before this change — so the local merge bought nothing and
carried the drift risk of a hand-rolled parser that has to round-trip every
field the backends read.

Drop it: All Green now sorts each cell's already-patched `lcov.info` into its
integration's directory and uploads the directory under one flag. Reports pass
through byte-for-byte, so the `getLineCoverage` patch the producers baked in
survives without needing a patched istanbul in All Green's sparse checkout, and
a future LCOV field needs no parser change.

Also drop the unused `CODECOV_TOKEN` env — the repo uploads tokenless and a
token would have to be an org secret that fork PRs cannot read.
